pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "h2mare"
version = "0.1.0"
description = "Downloader and data management tools for climate and ocean datasets."
authors = [
    { name = "Hugo Parra", email = "h2ugo.parra@pm.me" }
]
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["ocean", "climate", "geospatial", "zarr", "netcdf", "cmems", "copernicus", "era5"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "cartopy>=0.23.0",
    "cdsapi>=0.7.0",
    "cfgrib>=0.9.10",
    "copernicusmarine>=2.0.0",
    "dask>=2024.1.0",
    "duckdb>=1.0.0",
    "eccodes>=1.7.0",
    "ephem>=4.1.0",
    "geopandas>=1.0.0",
    "global_land_mask>=1.0.0",
    "ipython>=8.0.0",
    "loguru>=0.7.0",
    "matplotlib>=3.8.0",
    "msgspec>=0.18.0",
    "netCDF4>=1.6.0",
    "plotly>=5.18.0",
    "polars>=1.0.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
    "rioxarray>=0.17.0",
    "scipy>=1.13.0",
    "tqdm>=4.66.0",
    "typer>=0.12.0",
    "xarray>=2024.1.0",
]

[project.urls]
Repository = "https://github.com/h2ugoparra/h2gis"

[project.scripts]
h2mare = "h2mare.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["h2mare*"]

[tool.uv]
# sparse>=0.15 breaks copernicusmarine.subset
constraint-dependencies = ["sparse==0.14.0"]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "ipykernel>=7.2.0",
    "isort>=8.0.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "tox>=4.30.3",
    "tox-uv>=1.28.1",
]

[tool.isort]
profile = "black"
skip_glob = ["*.ipynb"]

[tool.tox]
legacy_tox_ini = """
[tox]
env_list = py311, py312, py313
isolated_build = true

[testenv]
runner = uv-venv-lock-runner
extras = dev
commands =
    pytest tests/ -q
"""
